# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Tender:
    properties:
      createdAt:
        type: string
      creatorUsername:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      organizationID:
        type: integer
      serviceType:
        type: string
      status:
        $ref: '#/definitions/models.TenderStatus'
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  models.TenderStatus:
    enum:
    - CREATED
    - PUBLISHED
    - CLOSED
    type: string
    x-enum-varnames:
    - CREATED
    - PUBLISHED
    - CLOSED
host: localhost:8080
info:
  contact: {}
  description: API для управления тендерами и предложениями
  title: Tender API
  version: "1.0"
paths:
...
  /tenders:
    get:
      consumes:
      - application/json
      description: Возвращает список всех тендеров с возможностью фильтрации по типу
        услуг.
      parameters:
      - description: Тип услуг для фильтрации тендеров
        in: query
        name: serviceType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список тендеров
          schema:
            items:
              $ref: '#/definitions/models.Tender'
            type: array
        "500":
          description: Ошибка загрузки тендеров
          schema:
            type: string
      summary: Получение списка тендеров
      tags:
      - Tenders
  /tenders/{tenderId}/edit:
    patch:
      consumes:
      - application/json
      description: Обновляет данные тендера (имя, описание, тип услуг) по его ID,
        если пользователь имеет права.
      parameters:
      - description: ID тендера
        in: path
        name: tenderId
        required: true
        type: integer
      - description: Имя пользователя, инициирующего изменение
        in: query
        name: username
        required: true
        type: string
      - description: Данные для обновления тендера (имя, описание, тип услуг)
        in: body
        name: tender
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный тендер
          schema:
            $ref: '#/definitions/models.Tender'
        "400":
          description: Неверные данные или ID тендера
          schema:
            type: string
        "403":
          description: Нет прав на редактирование тендера
          schema:
            type: string
        "404":
          description: Тендер или пользователь не найдены
          schema:
            type: string
        "500":
          description: Ошибка обновления тендера
          schema:
            type: string
      summary: Редактирование тендера
      tags:
      - Tenders
  /tenders/{tenderId}/rollback/{version}:
    put:
      consumes:
      - application/json
      description: Откатывает тендер к указанной версии на основании прав пользователя
        и статуса тендера. Откат невозможен, если тендер уже закрыт.
      parameters:
      - description: ID тендера
        in: path
        name: tenderId
        required: true
        type: integer
      - description: Версия тендера, к которой необходимо откатиться
        in: path
        name: version
        required: true
        type: integer
      - description: Имя пользователя, инициирующего откат
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Откатанный тендер
          schema:
            $ref: '#/definitions/models.Tender'
        "400":
          description: Неправильный ID тендера или версия
          schema:
            type: string
        "403":
          description: Нет прав на откат тендера
          schema:
            type: string
        "404":
          description: Тендер или версия не найдены
          schema:
            type: string
        "500":
          description: Ошибка сохранения откатанного тендера
          schema:
            type: string
      summary: Откат тендера к версии
      tags:
      - Tenders
  /tenders/{tenderId}/status:
    get:
      consumes:
      - application/json
      description: Возвращает статус тендера, если пользователь имеет права на просмотр
        статуса.
      parameters:
      - description: ID тендера
        in: path
        name: tenderId
        required: true
        type: integer
      - description: Имя пользователя, запрашивающего статус тендера
        in: query
        name: username
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Статус тендера
          schema:
            type: string
        "400":
          description: Неправильный ID тендера или никнейм не введен
          schema:
            type: string
        "403":
          description: Пользователь не является ответственным за тендер
          schema:
            type: string
        "404":
          description: Тендер или пользователь не найдены
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получение статуса тендера
      tags:
      - Tenders
    put:
      consumes:
      - application/json
      description: Позволяет изменить статус тендера на "publish" или "close", если
        пользователь имеет права доступа.
      parameters:
      - description: ID тендера
        in: path
        name: tenderId
        required: true
        type: integer
      - description: Имя пользователя, изменяющего статус
        in: query
        name: username
        required: true
        type: string
      - description: Новый статус тендера ('publish' или 'close')
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешно обновленный тендер
          schema:
            $ref: '#/definitions/models.Tender'
        "400":
          description: Неверный ID тендера или неправильный статус
          schema:
            type: string
        "403":
          description: Нет прав для изменения статуса
          schema:
            type: string
        "404":
          description: Тендер или пользователь не найдены
          schema:
            type: string
        "500":
          description: Ошибка обновления тендера
          schema:
            type: string
      summary: Изменение статуса тендера
      tags:
      - Tenders
  /tenders/my:
    get:
      consumes:
      - application/json
      description: Возвращает список всех тендеров, созданных пользователем по его
        имени.
      parameters:
      - description: Имя пользователя, создавшего тендеры
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список тендеров пользователя
          schema:
            items:
              $ref: '#/definitions/models.Tender'
            type: array
        "500":
          description: Ошибка поиска тендеров
          schema:
            type: string
      summary: Получение тендеров пользователя
      tags:
      - Tenders
  /tenders/new:
    post:
      consumes:
      - application/json
      description: Создает новый тендер, декодируя данные из тела запроса и сохраняя
        их в базе данных.
      parameters:
      - description: Данные для создания тендера
        in: body
        name: tender
        required: true
        schema:
          $ref: '#/definitions/models.Tender'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно созданный тендер
          schema:
            $ref: '#/definitions/models.Tender'
        "400":
          description: Неверные данные для создания тендера
          schema:
            type: string
        "500":
          description: Ошибка сохранения тендера в базе данных
          schema:
            type: string
      summary: Создание нового тендера
      tags:
      - Tenders
swagger: "2.0"